pyproject.toml
[project]
name = "sentinel"
version = "0.0.1"
description = "Sentinel — Langfuse LLMOps Agent (Deep Agents 기반)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.4.4",
    "langchain>=1.2",
    "langchain-openai>=1.1.10",
    "langfuse>=2.0",
    "langgraph>=1.0",
    "langgraph-checkpoint-sqlite>=1.0",
    "python-dotenv>=1.2.2",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "apscheduler>=3.10",
    "httpx>=0.28",
]

[project.optional-dependencies]
anthropic = ["langchain-anthropic>=0.3"]
gemini = ["langchain-google-genai>=2.1"]
ollama = ["langchain-ollama>=0.3"]
all-providers = [
    "langchain-anthropic>=0.3",
    "langchain-google-genai>=2.1",
    "langchain-ollama>=0.3",
]

[tool.setuptools.packages.find]
include = ["sentinel*"]

[project.scripts]
sentinel = "sentinel.cli:main"
